created Worksheets based on subcontrollers

+++ TaglistCreatorFromIGS/CreateTagListFromIGS.cs
@@ -23,7 +23,7 @@ namespace TaglistCreatorFromIGS
         private string excelFileFullPath = ""; // this is the full path with the excel file name
         Excel.Workbook xlWorkBook = null;
         Excel.Sheets worksheets = null;
-        Excel.Worksheet xlSht = null;
+        Excel.Worksheet xlActiveSheet = null;
 
         /*
          * this is the constructor
@@ -58,12 +58,33 @@ namespace TaglistCreatorFromIGS
         */
         private void createExcelFile()
         {
+            try
+            { 
             this.excelFileName = Path.GetFileNameWithoutExtension(this.fullPathToCSVDocument);
 
             string sourcePath = Path.GetDirectoryName(this.fullPathToCSVDocument); // this is directory of where the .csv file is located
             excelFileFullPath = System.IO.Path.Combine(sourcePath, excelFileName + ".xlsx");
-            System.IO.File.WriteAllBytes(excelFileFullPath, Properties.Resources.StandardTagList);  // this is used to copy the excel file in this projects resources and create a copy in the folder where the csv is located
+
+
+            if (!IsFileLocked(new FileInfo(excelFileFullPath)))
+                {
+                    System.IO.File.WriteAllBytes(excelFileFullPath, Properties.Resources.StandardTagList);  // this is used to copy the excel file in this projects resources and create a copy in the folder where the csv is located
+                }
+
+            else
+                {
+                    //throw();
+                    throw new Exception(String.Format(@"IOException: The file {0} is currently being used. {1}Follow the following steps: 
+                                                                    {1} 1. Delete the file 
+                                                                    {1} 2. If unable to delete; Close all programs using this file 
+                                                                    {1} 3. Rerun this program", excelFileFullPath,Environment.NewLine));
+                }
             
+            }
+            catch(Exception ex)
+            {
+                MessageBox.Show(ex.Message.ToString());
+            }
         }
 
 
@@ -80,29 +101,57 @@ namespace TaglistCreatorFromIGS
                 System.Windows.Forms.Application.Exit();
 
             }
-
+            xlApp.DisplayAlerts = false;
             xlWorkBook = xlApp.Workbooks.Open(excelFileFullPath, 0, false, 5, "", "", false, Microsoft.Office.Interop.Excel.XlPlatform.xlWindows, "", true, false, 0, true, false, false);
             worksheets = xlWorkBook.Worksheets;
-            xlSht = xlWorkBook.Sheets[1];
+            xlActiveSheet = xlWorkBook.Sheets[1];
         }
 
-        // this method is used to create worksheets
-        public void createWorksheets(string xlSheetName)
+        // this method is used to create the number of worksheets based on how many unique subcontroller files there are
+        // this method also changes the name of the worksheet based on the subcontroller files. 
+        private void createWorksheets()
         {
 
-            xlApp.DisplayAlerts = false;
+            // this section of the code is used to create X number of excel worksheet based on how many unique subcontrollers there are. 
+
+            //int uniqueSubControllerCount = csvDataTable.Count;
+
+            //for(int i = 0; i < uniqueSubControllerCount-1; i++) // reason for -1 is because the template already has 1 worksheet
+            //{
+            //    xlSht.Copy(Type.Missing, xlWorkBook.Sheets[xlWorkBook.Sheets.Count]);
+
+            //}
+
+
+
+
             
-            xlSht.Copy(Type.Missing, xlWorkBook.Sheets[xlWorkBook.Sheets.Count]);
-            xlWorkBook.Sheets[xlWorkBook.Sheets.Count].name = xlSheetName;
+            //xlWorkBook.Sheets[xlWorkBook.Sheets.Count].name = xlSheetName;
 
+            // this foreach loop is used to change the name of the worksheet to the name of the subcontrollers 
+
+            foreach (KeyValuePair<string, List<ParameterInfo>> entry in csvDataTable)
+            {
+                Debug.WriteLine(entry.Key.ToString() +" " + xlWorkBook.Sheets.Count.ToString());
+                xlActiveSheet.Copy(Type.Missing, xlWorkBook.Sheets[xlWorkBook.Sheets.Count]);
+                xlWorkBook.Sheets[xlWorkBook.Sheets.Count].name = entry.Key;
+            }
+
+            xlWorkBook.Sheets[1].delete(); // this is used to delete the first worksheet in the excel file that is marked as template
 
         }
 
-
-        public void writeDataToExcelWorksheets(string worksheetName, List<string> stringList)
+        /*
+         * Method: writeDataToExcelWorksheets
+         * Used to write the stringList list to the worksheet that is identified by worksheetName within the workbook object
+         */
+        private void writeDataToExcelWorksheets(Excel.Workbook xlWB, string worksheetName, List<ParameterInfo> stringList)
 
         {
+            xlActiveSheet =  xlWB.Sheets[worksheetName];
+            xlActiveSheet.Activate();
 
+            xlActiveSheet.Cells[3, 1] = worksheetName;
 
         }
 
@@ -122,6 +171,19 @@ namespace TaglistCreatorFromIGS
             xlWorkBook.Close();
 
             releaseObject(xlSht);
+            releaseObject(worksheets);
+            releaseObject(xlWorkBook);
+            releaseObject(xlApp);
+        }
+
+        // this is the overloaded method without any parameters
+        // used to close the class objects
+        private void saveCloseExcelFile()
+        {
+            xlWorkBook.Save();
+            xlWorkBook.Close();
+
+            releaseObject(xlActiveSheet);
             releaseObject(worksheets);
             releaseObject(xlWorkBook);
             releaseObject(xlApp);
@@ -148,14 +210,39 @@ namespace TaglistCreatorFromIGS
             }
         }
 
+        // this method is used to check if the file is currently being used.
+        protected virtual bool IsFileLocked(FileInfo file)
+        {
+            FileStream stream = null;
 
+            try
+            {
+                stream = file.Open(FileMode.Open, FileAccess.Read, FileShare.None);
+            }
+            catch (IOException)
+            {
+                //the file is unavailable because it is:
+                //still being written to
+                //or being processed by another thread
+                //or does not exist (has already been processed)
+                return true;
+            }
+            finally
+            {
+                if (stream != null)
+                    stream.Close();
+            }
+
+            //file is not locked
+            return false;
+        }
 
 
         public void readCSVFile()
         {
             //---------------------------------------------------------------------------
 
-            string csvFilePath = @"C:\Users\212478881\Desktop\TestCSV Folder\011110TestSiteFullIGSDriver.csv";
+            fullPathToCSVDocument = @"C:\Users\212478881\Desktop\TestCSV Folder\011110TestSiteFullIGSDriver.csv";
 
 
             //---------------------------------------------------------------------------
@@ -173,7 +260,7 @@ namespace TaglistCreatorFromIGS
             // Key: Sub Controller File name For example CMN, ZW1
             // Value: List of ParameterInfo object which hold the following parameters: Tag Name, Address, DataType
 
-            Dictionary<string, List<ParameterInfo>> csvDataTable = new Dictionary<string, List<ParameterInfo>>();
+            csvDataTable = new Dictionary<string, List<ParameterInfo>>();
 
             // this string is used to add entries to the dictionary using this as a key. 
             string currentSubControllerFile;
@@ -182,7 +269,7 @@ namespace TaglistCreatorFromIGS
             string currentAddress;
             string currentDataType;
 
-            using (TextFieldParser parser = new TextFieldParser(csvFilePath))
+            using (TextFieldParser parser = new TextFieldParser(fullPathToCSVDocument))
             {
                 parser.TextFieldType = FieldType.Delimited;
                 parser.SetDelimiters(",");
@@ -222,7 +309,7 @@ namespace TaglistCreatorFromIGS
                         csvDataTable.Add(currentSubControllerFile, newParameterList);
                     }
 
-
+                    
                 }
 
                 Debug.WriteLine("test");
@@ -260,10 +347,13 @@ namespace TaglistCreatorFromIGS
 
         public void generateTagList()
         {
+            readCSVFile();
             createExcelFile();
             initializeExcel();
-
-
+            createWorksheets();
+            string test = csvDataTable.Keys.First();
+            writeDataToExcelWorksheets(this.xlWorkBook, test, csvDataTable[test]);
+            saveCloseExcelFile();
         }
